pyproject.toml
[project]
name = "crypto-market-neutral-platform"
version = "1.0.0"
description = "Market-neutral automation platform for KRW premium and cross-exchange basis workflows"
requires-python = ">=3.9"
dependencies = [
    "ccxt>=4.0.0",
    "gate-api>=6.104.0",
    "PyJWT>=2.8.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[project.scripts]
market-neutral = "overseas_exchange_hedge.cli:main"
market-neutral-korea-entry = "overseas_exchange_hedge.cli:market_neutral_korea_entry"
market-neutral-overseas-unwind = "overseas_exchange_hedge.cli:market_neutral_overseas_unwind"
market-neutral-overseas-entry-auto = "overseas_exchange_hedge.cli:market_neutral_overseas_entry_auto"
market-neutral-overseas-entry-manual = "overseas_exchange_hedge.cli:market_neutral_overseas_entry_manual"
market-neutral-korea-exit = "overseas_exchange_hedge.cli:market_neutral_korea_exit"
hedge = "overseas_exchange_hedge.cli:main"
hedge-exit = "overseas_exchange_hedge.korea.exit.app:main"
hedge-redflag = "overseas_exchange_hedge.korea.redflag.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/overseas_exchange_hedge"]

[tool.uv]
dev-dependencies = [
    "mypy>=1.10.0",
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
    "types-requests>=2.31.0.20240914",
]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 120
src = ["src"]
extend-exclude = ["legacy-imports", "runtime"]

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
warn_unused_ignores = true
no_implicit_optional = true
check_untyped_defs = true
mypy_path = ["src"]
exclude = "(^runtime/)"
